+++ CSharpCADTests/src/unit_test_data/ExtrudeLinearTwistExp2.cs
@@ -5,144 +5,144 @@ using NUnit.Framework;
 public static partial class UnitTestData {
   public static List<List<Vec3>> ExtrudeLinearTwistExp2 = new List<List<Vec3>> {
       new List<Vec3> {
-        new Vec3((double)(5), (double)(-5), (double)(0)),
-        new Vec3((double)(5), (double)(5), (double)(0)),
-        new Vec3((double)(1.830127018922194), (double)(6.830127018922193), (double)(5))
-    },
+        new Vec3(5, -5, 0),
+        new Vec3(5, 5, 0),
+        new Vec3(1.830127018922194, 6.830127018922193, 5)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(5), (double)(-5), (double)(0)),
-        new Vec3((double)(1.830127018922194), (double)(6.830127018922193), (double)(5)),
-        new Vec3((double)(6.830127018922193), (double)(-1.830127018922194), (double)(5))
-    },
+        new Vec3(5, -5, 0),
+        new Vec3(1.830127018922194, 6.830127018922193, 5),
+        new Vec3(6.830127018922193, -1.830127018922194, 5)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(5), (double)(5), (double)(0)),
-        new Vec3((double)(-5), (double)(5), (double)(0)),
-        new Vec3((double)(-6.830127018922193), (double)(1.830127018922194), (double)(5))
-    },
+        new Vec3(5, 5, 0),
+        new Vec3(-5, 5, 0),
+        new Vec3(-6.830127018922193, 1.830127018922194, 5)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(5), (double)(5), (double)(0)),
-        new Vec3((double)(-6.830127018922193), (double)(1.830127018922194), (double)(5)),
-        new Vec3((double)(1.830127018922194), (double)(6.830127018922193), (double)(5))
-    },
+        new Vec3(5, 5, 0),
+        new Vec3(-6.830127018922193, 1.830127018922194, 5),
+        new Vec3(1.830127018922194, 6.830127018922193, 5)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-5), (double)(5), (double)(0)),
-        new Vec3((double)(-5), (double)(-5), (double)(0)),
-        new Vec3((double)(-1.830127018922194), (double)(-6.830127018922193), (double)(5))
-    },
+        new Vec3(-5, 5, 0),
+        new Vec3(-5, -5, 0),
+        new Vec3(-1.830127018922194, -6.830127018922193, 5)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-5), (double)(5), (double)(0)),
-        new Vec3((double)(-1.830127018922194), (double)(-6.830127018922193), (double)(5)),
-        new Vec3((double)(-6.830127018922193), (double)(1.830127018922194), (double)(5))
-    },
+        new Vec3(-5, 5, 0),
+        new Vec3(-1.830127018922194, -6.830127018922193, 5),
+        new Vec3(-6.830127018922193, 1.830127018922194, 5)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-5), (double)(-5), (double)(0)),
-        new Vec3((double)(5), (double)(-5), (double)(0)),
-        new Vec3((double)(6.830127018922193), (double)(-1.830127018922194), (double)(5))
-    },
+        new Vec3(-5, -5, 0),
+        new Vec3(5, -5, 0),
+        new Vec3(6.830127018922193, -1.830127018922194, 5)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-5), (double)(-5), (double)(0)),
-        new Vec3((double)(6.830127018922193), (double)(-1.830127018922194), (double)(5)),
-        new Vec3((double)(-1.830127018922194), (double)(-6.830127018922193), (double)(5))
-    },
+        new Vec3(-5, -5, 0),
+        new Vec3(6.830127018922193, -1.830127018922194, 5),
+        new Vec3(-1.830127018922194, -6.830127018922193, 5)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(6.830127018922193), (double)(-1.830127018922194), (double)(5)),
-        new Vec3((double)(1.830127018922194), (double)(6.830127018922193), (double)(5)),
-        new Vec3((double)(-1.8301270189221923), (double)(6.830127018922193), (double)(10))
-    },
+        new Vec3(6.830127018922193, -1.830127018922194, 5),
+        new Vec3(1.830127018922194, 6.830127018922193, 5),
+        new Vec3(-1.8301270189221923, 6.830127018922193, 10)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(6.830127018922193), (double)(-1.830127018922194), (double)(5)),
-        new Vec3((double)(-1.8301270189221923), (double)(6.830127018922193), (double)(10)),
-        new Vec3((double)(6.830127018922193), (double)(1.8301270189221923), (double)(10))
-    },
+        new Vec3(6.830127018922193, -1.830127018922194, 5),
+        new Vec3(-1.8301270189221923, 6.830127018922193, 10),
+        new Vec3(6.830127018922193, 1.8301270189221923, 10)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(1.830127018922194), (double)(6.830127018922193), (double)(5)),
-        new Vec3((double)(-6.830127018922193), (double)(1.830127018922194), (double)(5)),
-        new Vec3((double)(-6.830127018922193), (double)(-1.8301270189221923), (double)(10))
-    },
+        new Vec3(1.830127018922194, 6.830127018922193, 5),
+        new Vec3(-6.830127018922193, 1.830127018922194, 5),
+        new Vec3(-6.830127018922193, -1.8301270189221923, 10)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(1.830127018922194), (double)(6.830127018922193), (double)(5)),
-        new Vec3((double)(-6.830127018922193), (double)(-1.8301270189221923), (double)(10)),
-        new Vec3((double)(-1.8301270189221923), (double)(6.830127018922193), (double)(10))
-    },
+        new Vec3(1.830127018922194, 6.830127018922193, 5),
+        new Vec3(-6.830127018922193, -1.8301270189221923, 10),
+        new Vec3(-1.8301270189221923, 6.830127018922193, 10)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-6.830127018922193), (double)(1.830127018922194), (double)(5)),
-        new Vec3((double)(-1.830127018922194), (double)(-6.830127018922193), (double)(5)),
-        new Vec3((double)(1.8301270189221923), (double)(-6.830127018922193), (double)(10))
-    },
+        new Vec3(-6.830127018922193, 1.830127018922194, 5),
+        new Vec3(-1.830127018922194, -6.830127018922193, 5),
+        new Vec3(1.8301270189221923, -6.830127018922193, 10)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-6.830127018922193), (double)(1.830127018922194), (double)(5)),
-        new Vec3((double)(1.8301270189221923), (double)(-6.830127018922193), (double)(10)),
-        new Vec3((double)(-6.830127018922193), (double)(-1.8301270189221923), (double)(10))
-    },
+        new Vec3(-6.830127018922193, 1.830127018922194, 5),
+        new Vec3(1.8301270189221923, -6.830127018922193, 10),
+        new Vec3(-6.830127018922193, -1.8301270189221923, 10)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-1.830127018922194), (double)(-6.830127018922193), (double)(5)),
-        new Vec3((double)(6.830127018922193), (double)(-1.830127018922194), (double)(5)),
-        new Vec3((double)(6.830127018922193), (double)(1.8301270189221923), (double)(10))
-    },
+        new Vec3(-1.830127018922194, -6.830127018922193, 5),
+        new Vec3(6.830127018922193, -1.830127018922194, 5),
+        new Vec3(6.830127018922193, 1.8301270189221923, 10)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-1.830127018922194), (double)(-6.830127018922193), (double)(5)),
-        new Vec3((double)(6.830127018922193), (double)(1.8301270189221923), (double)(10)),
-        new Vec3((double)(1.8301270189221923), (double)(-6.830127018922193), (double)(10))
-    },
+        new Vec3(-1.830127018922194, -6.830127018922193, 5),
+        new Vec3(6.830127018922193, 1.8301270189221923, 10),
+        new Vec3(1.8301270189221923, -6.830127018922193, 10)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(6.830127018922193), (double)(1.8301270189221923), (double)(10)),
-        new Vec3((double)(-1.8301270189221923), (double)(6.830127018922193), (double)(10)),
-        new Vec3((double)(-5), (double)(5), (double)(15))
-    },
+        new Vec3(6.830127018922193, 1.8301270189221923, 10),
+        new Vec3(-1.8301270189221923, 6.830127018922193, 10),
+        new Vec3(-5, 5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(6.830127018922193), (double)(1.8301270189221923), (double)(10)),
-        new Vec3((double)(-5), (double)(5), (double)(15)),
-        new Vec3((double)(5), (double)(5), (double)(15))
-    },
+        new Vec3(6.830127018922193, 1.8301270189221923, 10),
+        new Vec3(-5, 5, 15),
+        new Vec3(5, 5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-1.8301270189221923), (double)(6.830127018922193), (double)(10)),
-        new Vec3((double)(-6.830127018922193), (double)(-1.8301270189221923), (double)(10)),
-        new Vec3((double)(-5), (double)(-5), (double)(15))
-    },
+        new Vec3(-1.8301270189221923, 6.830127018922193, 10),
+        new Vec3(-6.830127018922193, -1.8301270189221923, 10),
+        new Vec3(-5, -5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-1.8301270189221923), (double)(6.830127018922193), (double)(10)),
-        new Vec3((double)(-5), (double)(-5), (double)(15)),
-        new Vec3((double)(-5), (double)(5), (double)(15))
-    },
+        new Vec3(-1.8301270189221923, 6.830127018922193, 10),
+        new Vec3(-5, -5, 15),
+        new Vec3(-5, 5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-6.830127018922193), (double)(-1.8301270189221923), (double)(10)),
-        new Vec3((double)(1.8301270189221923), (double)(-6.830127018922193), (double)(10)),
-        new Vec3((double)(5), (double)(-5), (double)(15))
-    },
+        new Vec3(-6.830127018922193, -1.8301270189221923, 10),
+        new Vec3(1.8301270189221923, -6.830127018922193, 10),
+        new Vec3(5, -5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-6.830127018922193), (double)(-1.8301270189221923), (double)(10)),
-        new Vec3((double)(5), (double)(-5), (double)(15)),
-        new Vec3((double)(-5), (double)(-5), (double)(15))
-    },
+        new Vec3(-6.830127018922193, -1.8301270189221923, 10),
+        new Vec3(5, -5, 15),
+        new Vec3(-5, -5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(1.8301270189221923), (double)(-6.830127018922193), (double)(10)),
-        new Vec3((double)(6.830127018922193), (double)(1.8301270189221923), (double)(10)),
-        new Vec3((double)(5), (double)(5), (double)(15))
-    },
+        new Vec3(1.8301270189221923, -6.830127018922193, 10),
+        new Vec3(6.830127018922193, 1.8301270189221923, 10),
+        new Vec3(5, 5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(1.8301270189221923), (double)(-6.830127018922193), (double)(10)),
-        new Vec3((double)(5), (double)(5), (double)(15)),
-        new Vec3((double)(5), (double)(-5), (double)(15))
-    },
+        new Vec3(1.8301270189221923, -6.830127018922193, 10),
+        new Vec3(5, 5, 15),
+        new Vec3(5, -5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(5), (double)(-5), (double)(15)),
-        new Vec3((double)(5), (double)(5), (double)(15)),
-        new Vec3((double)(-5), (double)(5), (double)(15))
-    },
+        new Vec3(5, -5, 15),
+        new Vec3(5, 5, 15),
+        new Vec3(-5, 5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-5), (double)(5), (double)(15)),
-        new Vec3((double)(-5), (double)(-5), (double)(15)),
-        new Vec3((double)(5), (double)(-5), (double)(15))
-    },
+        new Vec3(-5, 5, 15),
+        new Vec3(-5, -5, 15),
+        new Vec3(5, -5, 15)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(5), (double)(5), (double)(0)),
-        new Vec3((double)(5), (double)(-5), (double)(0)),
-        new Vec3((double)(-5), (double)(-5), (double)(0))
-    },
+        new Vec3(5, 5, 0),
+        new Vec3(5, -5, 0),
+        new Vec3(-5, -5, 0)
+    }, 
       new List<Vec3> {
-        new Vec3((double)(-5), (double)(-5), (double)(0)),
-        new Vec3((double)(-5), (double)(5), (double)(0)),
-        new Vec3((double)(5), (double)(5), (double)(0))
+        new Vec3(-5, -5, 0),
+        new Vec3(-5, 5, 0),
+        new Vec3(5, 5, 0)
     }
   };
 }
